Front

+++ src/Controller/TaskController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Task;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -13,40 +14,110 @@ use Symfony\Component\Routing\Annotation\Route;
 class TaskController extends AbstractController
 {
     /**
-     * @Route("/task", name="task_list")
+     * @Route("/task", name="task_list_todo")
      * @param TaskRepository $repo
      * @return Response
      */
-    public function index(TaskRepository $repo): Response
+    public function indexDone(TaskRepository $repo): Response
     {
-        $tasks = $repo->findAll();
+        $tasks = $repo->findBy(['isDone' => false, 'inProgress' => false], ['createdAt' => 'DESC']);
 
-        return $this->render('task/index.html.twig', [
-            'controller_name' => 'TaskController',
-            'tasks' => $tasks
-        ]);
+        return $this->render(
+            'default/index.html.twig',
+            [
+                'controller_name' => 'TaskController',
+                'tasks' => $tasks,
+            ]
+        );
 
     }
 
     /**
+     * @Route("/task/done", name="task_list_done")
+     * @param TaskRepository $repo
+     * @return Response
+     */
+    public function indexNotDone(TaskRepository $repo): Response
+    {
+        $tasks = $repo->findBy(['isDone' => true], ['createdAt' => 'DESC']);
+
+        return $this->render(
+            'default/index.html.twig',
+            [
+                'controller_name' => 'TaskController',
+                'tasks' => $tasks,
+            ]
+        );
+
+    }
+
+    /**
+     * @Route("/task/inprogress", name="task_list_progress")
+     * @param TaskRepository $repo
+     * @return Response
+     */
+    public function indexInProgress(TaskRepository $repo): Response
+    {
+        $tasks = $repo->findBy(['inProgress' => true], ['createdAt' => 'DESC']);
+
+        //dd($tasks);
+
+        return $this->render(
+            'default/index.html.twig',
+            [
+                'controller_name' => 'TaskController',
+                'tasks' => $tasks,
+            ]
+        );
+
+    }
+
+    /**
+     * @Route("/task/assign/{id}", name="task_assign")
+     * @return Response
+     */
+    public function assignTask(): Response
+    {
+        $user = $this->getUser();
+
+
+        return $this->render(
+            'task/index.html.twig',
+            [
+                'controller_name' => 'TaskController',
+                'tasks' => $tasks,
+            ]
+        );
+
+    }
+
+
+    /**
      * @Route("/tasks/create", name="task_create")
+     * @IsGranted("ROLE_USER")
      */
     public function createAction(Request $request)
     {
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
 
+        $user = $this->getUser();
+
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
+
+            $task->setUser($user)
+                ->setInProgress(false);
+
 
             $em->persist($task);
             $em->flush();
 
             $this->addFlash('success', 'La tâche a été bien été ajoutée.');
 
-            return $this->redirectToRoute('task_list');
+            return $this->redirectToRoute('homepage');
         }
 
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
@@ -66,13 +137,16 @@ class TaskController extends AbstractController
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
 
-            return $this->redirectToRoute('task_list');
+            return $this->redirectToRoute('homepage');
         }
 
-        return $this->render('task/edit.html.twig', [
-            'form' => $form->createView(),
-            'task' => $task,
-        ]);
+        return $this->render(
+            'task/edit.html.twig',
+            [
+                'form' => $form->createView(),
+                'task' => $task,
+            ]
+        );
     }
 
     /**
@@ -85,7 +159,7 @@ class TaskController extends AbstractController
 
         $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
 
-        return $this->redirectToRoute('task_list');
+        return $this->redirectToRoute('homepage');
     }
 
     /**
@@ -93,12 +167,22 @@ class TaskController extends AbstractController
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        $user = $this->getUser();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        if ($this->isGranted('ROLE_ADMIN') || $task->getUser() == $user) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
 
-        return $this->redirectToRoute('task_list');
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+
+            return $this->redirectToRoute('homepage');
+
+        }
+        $this->addFlash('error', 'Vous n\'avez pas le droit de supprimer cette tâche.');
+
+        return $this->redirectToRoute('homepage');
+
+
     }
 }
